--- components/sign-in-form.tsx
@@ -1,104 +1,0 @@
-import { Link } from 'expo-router';
-import * as React from 'react';
-import { type TextInput, View } from 'react-native';
-
-import { SocialConnections } from '@/components/social-connections';
-import { Button } from '@/components/ui/button';
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
-import { Input } from '@/components/ui/input';
-import { Label } from '@/components/ui/label';
-import { Separator } from '@/components/ui/separator';
-import { Text } from '@/components/ui/text';
-
-/**
- *
- */
-export function SignInForm() {
-	const passwordInputRef = React.useRef<TextInput>(null);
-
-	/**
-	 *
-	 */
-	function onEmailSubmitEditing() {
-		passwordInputRef.current?.focus();
-	}
-
-	/**
-	 *
-	 */
-	function onSubmit() {
-		// TODO: Submit form and navigate to protected screen if successful
-	}
-
-	return (
-		<View className="gap-6">
-			<Card className="border-border/0 shadow-none sm:border-border sm:shadow-sm sm:shadow-black/5">
-				<CardHeader>
-					<CardTitle className="text-center text-xl sm:text-left">Log in, Goach in!</CardTitle>
-					<CardDescription className="text-center sm:text-left">
-						Welcome back! Please log in to continue
-					</CardDescription>
-				</CardHeader>
-				<CardContent className="gap-6">
-					<View className="gap-6">
-						<View className="gap-1.5">
-							<Label htmlFor="email">Email</Label>
-							<Input
-								id="email"
-								placeholder="goach@mail.com"
-								keyboardType="email-address"
-								autoComplete="email"
-								autoCapitalize="none"
-								onSubmitEditing={onEmailSubmitEditing}
-								returnKeyType="next"
-								submitBehavior="submit"
-							/>
-						</View>
-						<View className="gap-1.5">
-							<View className="flex-row items-center">
-								<Label htmlFor="password">Password</Label>
-								<Button
-									variant="link"
-									size="sm"
-									className="ml-auto h-4 px-1 py-0 web:h-fit sm:h-4"
-									onPress={() => {
-										// TODO: Navigate to forgot password screen
-									}}
-								>
-									<Text className="font-normal leading-4">Forgot your password?</Text>
-								</Button>
-							</View>
-							<Input
-								ref={passwordInputRef}
-								id="password"
-								secureTextEntry
-								returnKeyType="send"
-								onSubmitEditing={onSubmit}
-							/>
-						</View>
-						<Button className="w-full" onPress={onSubmit}>
-							<Text>Continue</Text>
-						</Button>
-					</View>
-					<Text className="text-center text-sm">
-						No accounts? Guess you'll have to{' '}
-						<Link
-							href="/register"
-							className="text-green-400"
-							style={{ textDecorationLine: 'underline' }}
-						>
-							Register
-						</Link>{' '}
-						first heh.
-					</Text>
-					<View className="flex-row items-center">
-						<Separator className="flex-1" />
-						<Text className="px-4 text-sm text-muted-foreground">or</Text>
-						<Separator className="flex-1" />
-					</View>
-					<SocialConnections />
-				</CardContent>
-			</Card>
-		</View>
-	);
-}
