--- index.ts
@@ -1,6 +1,0 @@
-import { registerRootComponent } from 'expo';
-import App from './App';
-// RegisterRootComponent calls AppRegistry.registerComponent('main', () => App);
-// It also ensures that whether you load the app in Expo Go or in a native build,
-// The environment is set up appropriately
-registerRootComponent(App);
