+++ src/app/_layout.tsx
@@ -1,6 +1,6 @@
 import { PortalHost } from '@rn-primitives/portal';
 import { Stack } from 'expo-router';
-import '../global.css';
+import '../../global.css';
 /**
  *
  */
