+++ src/app/components/ui/card.tsx
@@ -1,5 +1,5 @@
-import { Text, TextClassContext } from '@/components/ui/text';
-import { cn } from '@/lib/utils';
+import { Text, TextClassContext } from '@/src/app/components/ui/text';
+import { cn } from '@/src/app/lib/utils';
 import { View, type ViewProps } from 'react-native';
 
 function Card({ className, ...props }: ViewProps & React.RefAttributes<View>) {
@@ -8,7 +8,7 @@ function Card({ className, ...props }: ViewProps & React.RefAttributes<View>) {
       <View
         className={cn(
           'bg-card border-border flex flex-col gap-6 rounded-xl border py-6 shadow-sm shadow-black/5',
-          className
+          className,
         )}
         {...props}
       />
