+++ src/app/components/ui/input.tsx
@@ -1,4 +1,4 @@
-import { cn } from '@/lib/utils';
+import { cn } from '@/src/app/lib/utils';
 import { Platform, TextInput, type TextInputProps } from 'react-native';
 
 function Input({
@@ -13,17 +13,17 @@ function Input({
         props.editable === false &&
           cn(
             'opacity-50',
-            Platform.select({ web: 'disabled:pointer-events-none disabled:cursor-not-allowed' })
+            Platform.select({ web: 'disabled:pointer-events-none disabled:cursor-not-allowed' }),
           ),
         Platform.select({
           web: cn(
             'placeholder:text-muted-foreground selection:bg-primary selection:text-primary-foreground outline-none transition-[color,box-shadow] md:text-sm',
             'focus-visible:border-ring focus-visible:ring-ring/50 focus-visible:ring-[3px]',
-            'aria-invalid:ring-destructive/20 dark:aria-invalid:ring-destructive/40 aria-invalid:border-destructive'
+            'aria-invalid:ring-destructive/20 dark:aria-invalid:ring-destructive/40 aria-invalid:border-destructive',
           ),
           native: 'placeholder:text-muted-foreground/50',
         }),
-        className
+        className,
       )}
       {...props}
     />
