+++ src/app/components/ui/native-only-animated-view.tsx
