+++ src/app/components/ui/switch.tsx
@@ -1,7 +1,11 @@
-import { cn } from '@/lib/utils';
 import * as SwitchPrimitives from '@rn-primitives/switch';
 import { Platform } from 'react-native';
 
+import { cn } from '@/src/app/lib/utils';
+
+/**
+ *
+ */
 function Switch({
   className,
   ...props
@@ -15,9 +19,10 @@ function Switch({
         }),
         props.checked ? 'bg-primary' : 'bg-input dark:bg-input/80',
         props.disabled && 'opacity-50',
-        className
+        className,
       )}
-      {...props}>
+      {...props}
+    >
       <SwitchPrimitives.Thumb
         className={cn(
           'bg-background size-4 rounded-full transition-transform',
@@ -26,7 +31,7 @@ function Switch({
           }),
           props.checked
             ? 'dark:bg-primary-foreground translate-x-3.5'
-            : 'dark:bg-foreground translate-x-0'
+            : 'dark:bg-foreground translate-x-0',
         )}
       />
     </SwitchPrimitives.Root>
