+++ src/app/components/ui/text.tsx
@@ -1,4 +1,4 @@
-import { cn } from '@/lib/utils';
+import { cn } from '@/src/app/lib/utils';
 import * as Slot from '@rn-primitives/slot';
 import { cva, type VariantProps } from 'class-variance-authority';
 import * as React from 'react';
@@ -9,7 +9,7 @@ const textVariants = cva(
     'text-foreground text-base',
     Platform.select({
       web: 'select-text',
-    })
+    }),
   ),
   {
     variants: {
@@ -17,18 +17,18 @@ const textVariants = cva(
         default: '',
         h1: cn(
           'text-center text-4xl font-extrabold tracking-tight',
-          Platform.select({ web: 'scroll-m-20 text-balance' })
+          Platform.select({ web: 'scroll-m-20 text-balance' }),
         ),
         h2: cn(
           'border-border border-b pb-2 text-3xl font-semibold tracking-tight',
-          Platform.select({ web: 'scroll-m-20 first:mt-0' })
+          Platform.select({ web: 'scroll-m-20 first:mt-0' }),
         ),
         h3: cn('text-2xl font-semibold tracking-tight', Platform.select({ web: 'scroll-m-20' })),
         h4: cn('text-xl font-semibold tracking-tight', Platform.select({ web: 'scroll-m-20' })),
         p: 'mt-3 leading-7 sm:mt-6',
         blockquote: 'mt-4 border-l-2 pl-3 italic sm:mt-6 sm:pl-6',
         code: cn(
-          'bg-muted relative rounded px-[0.3rem] py-[0.2rem] font-mono text-sm font-semibold'
+          'bg-muted relative rounded px-[0.3rem] py-[0.2rem] font-mono text-sm font-semibold',
         ),
         lead: 'text-muted-foreground text-xl',
         large: 'text-lg font-semibold',
@@ -39,7 +39,7 @@ const textVariants = cva(
     defaultVariants: {
       variant: 'default',
     },
-  }
+  },
 );
 
 type TextVariantProps = VariantProps<typeof textVariants>;
